Added details page to maps page functionality
When the users clicks on the Pin icon in the Details page of a post, it takes them to the map screen with the exact location of the post.
Although it works, it causes the views to stack on each other and allows for an infinite/constant stack (needs to be fixed later).
In addition, users can now only post if they are allowing location services.
TODO:
- Implement privacy functionalities when posting