Revised Account Deletion (no change)
I attempted to create reusable functions but in the end, it resulted in less than 10 lines of code saved and a not so scalable approach in case new data fields are eventually introduced. I decided to continue with this approach, even though the function is around 60 lines long. In case it expands from there, I will make sure to decompose the function into various small ones.
In addition, I concluded that using the synchronous methods for deletion was the best route in order to ensure that by the time the user is about to be deleted and signed out, no other data is also being deleted in a different thread and could cause an error.