Added External Menu Library + Delete Functionality
External library is used to create a dropdown menu in the settings so that the user can choose the specific data they want to delete.
Had to create specific delete functions for each data in order to allow these changes. In addition, an alert is shown when the deletion is complete.